+++ changelog.md
@@ -10,3 +10,6 @@ Changes to the lab will be details in this log file.
 #### Feb 04 2016
 - fixed second image link (lockemup)
 - fixed third image link (show)
+
+####Feb 05 2015
+- fixed fourth image link (dunno)